pyproject.toml
[project]
name = "z-register"
version = "0.0.1"
description = "tele: @ovftank"
requires-python = ">=3.10.19"
authors = [{ name = "ovftank", email = "ovftank@gmail.com" }]
dependencies = [
    "onnx==1.19.1",
    "onnxruntime==1.23.2",
    "pillow==12.0.0",
    "playwright==1.55.0",
    "playwright-stealth==2.0.0",
    "psutil==7.1.3",
    "pyinstaller==6.16.0",
    "python-ghost-cursor==0.1.1",
    "requests==2.32.5",
    "screeninfo==0.8.1",
]
